% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_estimates}
\alias{bootstrap_estimates}
\title{Function to replicate n_boot bootstrap samples and get bootstrap standard error}
\usage{
bootstrap_estimates(
  data,
  G_name = "G",
  V_name = "V",
  X_name = "X",
  Y_name = "Y",
  n_boot = 1000,
  est = c("gcomp_pop_estimand", "gcomp", "efficient_aipw", "efficient_tmle"),
  G_X_model = NULL,
  Y_X_model = NULL,
  family = "gaussian"
)
}
\arguments{
\item{data}{original data to bootstrap}

\item{G_name}{growth outcome variable name}

\item{V_name}{vaccination variable name}

\item{X_name}{covariate name(s)}

\item{Y_name}{infection variable name}

\item{n_boot}{number of bootstrap replicates}

\item{est}{character vector of names of estimators to use for growth effect}

\item{G_X_model}{optional specify model to be used for fitting growth on covariates, otherwise growth on all covariates}

\item{Y_X_model}{optional specify model to be used for fitting infection on covariates, otherwise infection on all covariates}

\item{family}{family for outcome model, defaults to gaussian for growth}
}
\value{
list containing bootstrap se and 95\% CI bounds for estimators specified in est
}
\description{
Function to replicate n_boot bootstrap samples and get bootstrap standard error
}
