% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_doomed.R
\name{do_aipw_doomed}
\alias{do_aipw_doomed}
\title{Function for efficient AIPW estimator in the doomed stratum}
\usage{
do_aipw_doomed(
  data,
  models,
  Y_name = "Y",
  Z_name = "Z",
  S_name = "S",
  return_se = FALSE
)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{Y_name}{name of growth outcome variable, default Y}

\item{Z_name}{name of vaccine treatment variable, default Z}

\item{S_name}{name of infection variable, default Y}

\item{return_se}{flag to return standard error, defualt FALSE}
}
\value{
AIPW estimate of growth effect in doomed stratum (+ standard error if return_se = TRUE)
}
\description{
Function for efficient AIPW estimator in the doomed stratum
}
