% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nat_inf.R
\name{do_aipw_nat_inf}
\alias{do_aipw_nat_inf}
\title{Function for efficient AIPW estimator}
\usage{
do_aipw_nat_inf(
  data,
  models,
  exclusion_restriction = FALSE,
  cross_world = TRUE,
  Y_name = "Y",
  Z_name = "Z",
  S_name = "S",
  X_name = "X",
  return_se = FALSE,
  two_part_model = FALSE
)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{exclusion_restriction}{boolean indicating whether an exclusion restriction is assumed}

\item{cross_world}{boolean indicating whether a cross-world independence assumption is assumed}

\item{Y_name}{name of growth outcome variable, default Y}

\item{Z_name}{name of vaccine treatment variable, default Z}

\item{S_name}{name of infection variable, default Y}

\item{return_se}{flag to return standard error, defualt FALSE}
}
\value{
AIPW estimate of growth effect in naturally infected strata (+ standard error if return_se = TRUE)
}
\description{
Function for efficient AIPW estimator
}
