% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pop.R
\name{do_aipw_pop}
\alias{do_aipw_pop}
\title{Function for AIPW of average counterfactual post-infection outcome marginally}
\usage{
do_aipw_pop(
  data,
  models,
  Z_name = "Z",
  X_name = c("X"),
  Y_name = "Y",
  return_se = TRUE,
  two_part_model = FALSE
)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{Z_name}{name of vaccine treatment variable, default Z}

\item{X_name}{character vector containing name(s) of covariates, default X}

\item{S_name}{character vector containing name(s) of covariates, default Y}
}
\value{
g-comp estimate of growth effect for population estimand
}
\description{
Function for AIPW of average counterfactual post-infection outcome marginally
}
