% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_tests.R
\name{do_chop_lump_test}
\alias{do_chop_lump_test}
\title{Function to do permutation test for chop-lump style test-statistics}
\usage{
do_chop_lump_test(
  data,
  G_name = "G",
  V_name = "V",
  Y_name = "Y",
  n_permutations = 10000
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{G_name}{growth outcome variable name}

\item{V_name}{vaccination variable name}

\item{Y_name}{infection variable name}

\item{n_permutations}{number of permutations to complete}
}
\value{
chop lump test statistic
}
\description{
Function to do permutation test for chop-lump style test-statistics
}
