% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{do_efficient_tmle}
\alias{do_efficient_tmle}
\title{Function for efficient TMLE estimator}
\usage{
do_efficient_tmle(
  data,
  models,
  G_name = "G",
  V_name = "V",
  Y_name = "Y",
  return_se = FALSE,
  max_iter = 10,
  tol = 1/(sqrt(dim(data)[1]) * log(dim(data)[1]))
)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{G_name}{name of growth outcome variable, default G}

\item{V_name}{name of vaccine treatment variable, default V}

\item{Y_name}{name of infection variable, default Y}

\item{return_se}{flag to return standard error, defualt FALSE}

\item{max_iter}{TODO}

\item{tol}{TOOD}
}
\value{
TMLE estimate of growth effect (+ standard error if return_se = TRUE)
}
\description{
Function for efficient TMLE estimator
}
