% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{do_gcomp}
\alias{do_gcomp}
\title{Function for g-computation of VE estimand}
\usage{
do_gcomp(data, models)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}
}
\value{
g-comp estimate of growth effect for VE estimand
}
\description{
Function for g-computation of VE estimand
}
