% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_doomed.R
\name{do_gcomp_doomed}
\alias{do_gcomp_doomed}
\title{Function for g-computation of counterfactual post-infection outcomes in the
doomed principal strata}
\usage{
do_gcomp_doomed(data, models)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}
}
\value{
g-comp estimate of growth effect in the doomed strata
}
\description{
Function for g-computation of counterfactual post-infection outcomes in the
doomed principal strata
}
