% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nat_inf.R
\name{do_gcomp_nat_inf}
\alias{do_gcomp_nat_inf}
\title{Function for g-computation of counterfactual post-infection outcomes in the
naturally infected principal strata}
\usage{
do_gcomp_nat_inf(
  data,
  models,
  Z_name = NULL,
  X_name = NULL,
  exclusion_restriction = FALSE
)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}
}
\value{
g-comp estimate of growth effect in the naturally infected strata
}
\description{
Function for g-computation of counterfactual post-infection outcomes in the
naturally infected principal strata
}
