% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pop.R
\name{do_gcomp_pop}
\alias{do_gcomp_pop}
\title{Function for g-computation of average counterfactual post-infection outcome marginally}
\usage{
do_gcomp_pop(data, models, Z_name = "Z", X_name = c("X"))
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{Z_name}{name of vaccine treatment variable, default Z}

\item{X_name}{character vector containing name(s) of covariates, default X}
}
\value{
g-comp estimate of growth effect for population estimand
}
\description{
Function for g-computation of average counterfactual post-infection outcome marginally
}
