% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{do_gcomp_short_term}
\alias{do_gcomp_short_term}
\title{Function for g-computation of short-term growth estimand}
\usage{
do_gcomp_short_term(data, models, V_name = "V", X_name = c("X"))
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{V_name}{name of vaccine treatment variable, default V}

\item{X_name}{character vector containing name(s) of covariates, default X}
}
\value{
g-comp estimate of growth effect for short term growth estimand
}
\description{
Function for g-computation of short-term growth estimand
}
