% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_doomed.R
\name{do_ipw_doomed}
\alias{do_ipw_doomed}
\title{Function for IPW of counterfactual post-infection outcomes in the
doomed principal stratum}
\usage{
do_ipw_doomed(data, models, S_name, Y_name, Z_name)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}
}
\value{
IPW estimate of growth effect in the doomed principal stratum
}
\description{
Function for IPW of counterfactual post-infection outcomes in the
doomed principal stratum
}
