% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nat_inf.R
\name{do_ipw_nat_inf}
\alias{do_ipw_nat_inf}
\title{Function for IPW of counterfactual post-infection outcomes in the
naturally infected principal strata}
\usage{
do_ipw_nat_inf(data, models, S_name, Y_name, Z_name)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{S_name}{TODO}

\item{Y_name}{TODO}

\item{Z_name}{TODO}
}
\value{
IPW estimate of growth effect in the naturally infected principal stratum
}
\description{
Function for IPW of counterfactual post-infection outcomes in the
naturally infected principal strata
}
