% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pop.R
\name{do_ipw_pop}
\alias{do_ipw_pop}
\title{Function for IPW of average counterfactual post-infection outcome marginally}
\usage{
do_ipw_pop(data, models, Z_name = "Z", Y_name = "Y")
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{Z_name}{name of vaccine treatment variable, default Z}

\item{Y_name}{character vector containing name(s) of post-infection outcome, default Y}
}
\value{
g-comp estimate of growth effect for population estimand
}
\description{
Function for IPW of average counterfactual post-infection outcome marginally
}
