% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{do_one_ci_sim}
\alias{do_one_ci_sim}
\title{Function to run simulation for single seed}
\usage{
do_one_ci_sim(
  params,
  n_boot = 1000,
  est = c("gcomp_pop_estimand", "gcomp", "efficient_aipw", "efficient_tmle", "choplump"),
  null_hypothesis_value = 0,
  alpha_level = 0.025
)
}
\arguments{
\item{params}{parameters used for data generation}

\item{n_boot}{number of bootstrap replicates}

\item{est}{character vector of names of estimators to use for growth effect}

\item{null_hypothesis_value}{value of null hypothesis for growth effect hypothesis testing, default 0}

\item{alpha_level}{one-sided alpha for hypothesis testing, default 0.025}
}
\value{
dataframe with parameters, point estimates, confidence intervals for specified estimators
}
\description{
Function to run simulation for single seed
}
