% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{do_sens_aipw}
\alias{do_sens_aipw}
\title{Function for efficient AIPW estimator for sensitivity analysis}
\usage{
do_sens_aipw(
  data,
  models,
  G_name = "G",
  V_name = "V",
  Y_name = "Y",
  epsilon = exp(seq(log(0.5), log(2), length = 50)),
  return_se = FALSE
)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{G_name}{name of growth outcome variable, default G}

\item{V_name}{name of vaccine treatment variable, default V}

\item{Y_name}{name of infection variable, default Y}

\item{return_se}{flag to return standard error, defualt FALSE}

\item{epislon}{a vector of values for the sensitivity parameter}
}
\value{
AIPW estimate of growth effect (+ standard error if return_se = TRUE)
}
\description{
Function for efficient AIPW estimator for sensitivity analysis
}
\examples{
data(provide)
models <- fit_models(
  data,
  G_name = "any_abx_wk52",
  V_name = "rotaarm",
  X_name = c("wk10_haz", "gender", "num_hh_sleep"),
  Y_name = "rotaepi",
  est = "efficient_aipw"
)

# debug(do_sens_aipw)
out <- do_sens_aipw(
  provide, models, 
  G_name = "any_abx_wk52",
  V_name = "rotaarm",
  Y_name = "rotaepi",
  return_se = TRUE
)

plot(out, se = TRUE, effect_type = "additive")
plot(out, se = TRUE, effect_type = "multiplicative")
}
