% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nat_inf.R
\name{do_sens_aipw_nat_inf}
\alias{do_sens_aipw_nat_inf}
\title{Function for efficient AIPW estimator for sensitivity analysis}
\usage{
do_sens_aipw_nat_inf(
  data,
  models,
  Y_name = "Y",
  Z_name = "Z",
  S_name = "S",
  epsilon = exp(seq(log(0.5), log(2), length = 49)),
  return_se = FALSE
)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{Y_name}{name of growth outcome variable, default Y}

\item{Z_name}{name of vaccine treatment variable, default Z}

\item{S_name}{name of infection variable, default Y}

\item{return_se}{flag to return standard error, defualt FALSE}

\item{epislon}{a vector of values for the sensitivity parameter}
}
\value{
AIPW estimate of growth effect (+ standard error if return_se = TRUE)
}
\description{
Function for efficient AIPW estimator for sensitivity analysis
}
\examples{
data(provide)
models <- fit_models(
  data,
  Y_name = "anS_abx_wk52",
  Z_name = "rotaarm",
  X_name = c("wk10_haz", "gender", "num_hh_sleep"),
  S_name = "rotaepi",
  est = "efficient_aipw"
)

# debug(do_sens_aipw)
out <- do_sens_aipw(
  provide, models, 
  Y_name = "anS_abx_wk52",
  Z_name = "rotaarm",
  S_name = "rotaepi",
  return_se = TRUE
)

plot(out, se = TRUE, effect_type = "additive")
plot(out, se = TRUE, effect_type = "multiplicative")
}
