% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nat_inf.R
\name{do_tmle_nat_inf}
\alias{do_tmle_nat_inf}
\title{Function for efficient TMLE estimator}
\usage{
do_tmle_nat_inf(
  data,
  models,
  Y_name = "Y",
  Z_name = "Z",
  S_name = "S",
  return_se = FALSE,
  max_iter = 10,
  tol = 1/(sqrt(dim(data)[1]) * log(dim(data)[1]))
)
}
\arguments{
\item{data}{dataset to predict on}

\item{models}{list of pre-fit models needed for estimation}

\item{Y_name}{name of growth outcome variable, default Y}

\item{Z_name}{name of vaccine treatment variable, default Z}

\item{S_name}{name of infection variable, default Y}

\item{return_se}{flag to return standard error, defualt FALSE}

\item{max_iter}{TODO}

\item{tol}{TOOD}
}
\value{
TMLE estimate of growth effect (+ standard error if return_se = TRUE)
}
\description{
Function for efficient TMLE estimator
}
