% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nat_inf.R
\name{do_unadj_nat_inf}
\alias{do_unadj_nat_inf}
\title{Function for covariate-unadjusted estimation of counterfactual post-infection outcomes in the
naturally infected principal strata}
\usage{
do_unadj_nat_inf(data, Z_name, Y_name, S_name)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{Z_name}{vaccination variable name}

\item{Y_name}{growth outcome variable name}

\item{S_name}{infection variable name}
}
\value{
unadjusted estimates of growth effect in the naturally infected strata
}
\description{
Function for covariate-unadjusted estimation of counterfactual post-infection outcomes in the
naturally infected principal strata
}
