% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{fit_ml_models}
\alias{fit_ml_models}
\title{Function to fit all SuperLearner models needed by various estimators}
\usage{
fit_ml_models(
  data,
  G_name = "G",
  V_name = "V",
  X_name = c("X"),
  Y_name = "Y",
  est = c("gcomp_pop_estimand", "gcomp", "efficient_aipw", "efficient_tmle"),
  G_V_X_library = c("SL.glm"),
  G_X_library = c("SL.glm"),
  Y_X_library = c("SL.glm"),
  family = "gaussian",
  v_folds = 3
)
}
\arguments{
\item{data}{dataset used to fit models}

\item{G_name}{name of growth outcome variable, default G}

\item{V_name}{name of vaccine treatment variable, default V}

\item{X_name}{character vector containing name(s) of covariates, default X}

\item{Y_name}{name of infection variable, default Y}

\item{est}{character vector of names of estimators to use for growth effect}

\item{G_X_library}{specify SuperLearner libraries to be used for fitting growth on covariates. Default GLM}

\item{Y_X_library}{specify SuperLearner libraries to be used for fitting infection on covariates. Default GLM}

\item{family}{family for outcome model, defaults to gaussian for growth}

\item{v_folds}{number of cross-validation folds to use in SuperLearner, default 3}
}
\value{
list with models needed for specified estimators
}
\description{
Function to fit all SuperLearner models needed by various estimators
}
