% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{fit_models}
\alias{fit_models}
\title{Function to fit all models needed by various estimators}
\usage{
fit_models(
  data,
  G_name = "G",
  V_name = "V",
  X_name = c("X"),
  Y_name = "Y",
  est = c("gcomp_pop_estimand", "gcomp", "efficient_aipw", "efficient_tmle"),
  G_X_model = NULL,
  Y_X_model = NULL,
  family = "gaussian"
)
}
\arguments{
\item{data}{dataset used to fit models}

\item{G_name}{name of growth outcome variable, default G}

\item{V_name}{name of vaccine treatment variable, default V}

\item{X_name}{character vector containing name(s) of covariates, default X}

\item{Y_name}{name of infection variable, default Y}

\item{est}{character vector of names of estimators to use for growth effect}

\item{G_X_model}{optional specify model to be used for fitting growth on covariates, otherwise growth on all covariates}

\item{Y_X_model}{optional specify model to be used for fitting infection on covariates, otherwise infection on all covariates}

\item{family}{family for outcome model, defaults to gaussian for growth}
}
\value{
list with models needed for specified estimators
}
\description{
Function to fit all models needed by various estimators
}
