% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_boot_se}
\alias{get_boot_se}
\title{Helper function to get SE and 95\% CI for additive and multiplicative effects - point estimates}
\usage{
get_boot_se(boot_estimates, estimand = "nat_inf", method = "gcomp")
}
\arguments{
\item{estimand}{estimand of interest}

\item{method}{method of interest}
}
\value{
dataframe with bootstrap standard error and confidence intervals on additive and multiplicative scales for point estimates
}
\description{
Helper function to get SE and 95\% CI for additive and multiplicative effects - point estimates
}
