% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_boot_se_sens}
\alias{get_boot_se_sens}
\title{Helper function to get SE and 95\% CI for additive and multiplicative effects - bounds}
\usage{
get_boot_se_sens(boot_estimates, estimand = "nat_inf", method = "sens")
}
\arguments{
\item{estimand}{estimand of interest}

\item{method}{method of interest}
}
\value{
dataframe with bootstrap standard error and confidence intervals on additive and multiplicative scales for each epsilon
}
\description{
Helper function to get SE and 95\% CI for additive and multiplicative effects - bounds
}
