% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_doomed.R
\name{get_bound_doomed}
\alias{get_bound_doomed}
\title{Function for bounds on doomed estimate without use of cross-world assumption}
\usage{
get_bound_doomed(
  data,
  Y_name = "Y",
  Z_name = "Z",
  S_name = "S",
  family = "gaussian"
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{Y_name}{growth outcome variable name}

\item{Z_name}{vaccination variable name}

\item{S_name}{infection variable name}

\item{family}{gaussian for continuous outcome, binomial for binary outcome}
}
\value{
list containing estimate of E\link{Y(1) | Y(0) = 1}, bounds on E\link{Y(0) | Y(0) = 1}, bounds on additive effect, bounds on multiplicative effect
}
\description{
Function for bounds on doomed estimate without use of cross-world assumption
}
