% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_tests.R
\name{get_chop_lump_statistic}
\alias{get_chop_lump_statistic}
\title{Function to get chop-lump style test-statistic}
\usage{
get_chop_lump_statistic(data, G_name = "G", V_name = "V", Y_name = "Y")
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{G_name}{growth outcome variable name}

\item{V_name}{vaccination variable name}

\item{Y_name}{infection variable name}
}
\value{
dataframe with chop-lump style test statistics for mean in vax, mean in placebo
}
\description{
Function to get chop-lump style test-statistic
}
