% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nat_inf.R
\name{get_cov_adj_bound_nat_inf}
\alias{get_cov_adj_bound_nat_inf}
\title{Function for covariate-adjusted bounds on naturally infected estimate}
\usage{
get_cov_adj_bound_nat_inf(
  data,
  X_name = "X",
  Y_name = "Y",
  Z_name = "Z",
  S_name = "S",
  family = "gaussian"
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{X_name}{discrete-valued covariate name}

\item{Y_name}{growth outcome variable name}

\item{Z_name}{vaccination variable name}

\item{S_name}{infection variable name}

\item{family}{gaussian for continuous outcome, binomial for binary outcome}
}
\value{
list containing estimate of E\link{Y(0) | Y(0) = 1}, bounds on E\link{Y(1) | Y(0) = 1}, bounds on additive effect, bounds on multiplicative effect
}
\description{
Function for covariate-adjusted bounds on naturally infected estimate
}
