% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_tests.R
\name{get_hudgens_stat}
\alias{get_hudgens_stat}
\title{Function for hudgens-style test statistic}
\usage{
get_hudgens_stat(
  data,
  G_name = "G",
  V_name = "V",
  Y_name = "Y",
  lower_bound = TRUE
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{G_name}{growth outcome variable name}

\item{V_name}{vaccination variable name}

\item{Y_name}{infection variable name}

\item{lower_bound}{A boolean. If TRUE, then adds the smallest growth measures
to the infected vaccinees thereby yielding a lower
bound on the effect of interest. If FALSE, then adds the largest
growth measures to the infected vacccinees thereby yielding an upper
bound on the effect of interest.}
}
\value{
Hudgens-style test statistic
}
\description{
Function for hudgens-style test statistic
}
