% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_tests.R
\name{get_hudgens_stat_doomed_new}
\alias{get_hudgens_stat_doomed_new}
\title{Function for hudgens-style test statistic - match description in the paper}
\usage{
get_hudgens_stat_doomed_new(
  data,
  G_name = "G",
  V_name = "V",
  Y_name = "Y",
  family = "gaussian"
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{G_name}{growth outcome variable name}

\item{V_name}{vaccination variable name}

\item{Y_name}{infection variable name}

\item{family}{gaussian for continuous outcome, binomial for binary}
}
\value{
Hudgens-style test statistic
}
\description{
Function for hudgens-style test statistic - match description in the paper
}
