% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_tests.R
\name{hudgens_test_new}
\alias{hudgens_test_new}
\title{Function for bootstrap test of Hudgens-style test statistic}
\usage{
hudgens_test_new(
  data,
  G_name = "G",
  V_name = "V",
  Y_name = "Y",
  family = "gaussian",
  n_boot = 1000,
  n_boot_try = n_boot * 10,
  effect_dir = "positive"
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{G_name}{growth outcome variable name}

\item{V_name}{vaccination variable name}

\item{Y_name}{infection variable name}

\item{family}{gaussian for continuous outcome, binomial for binary}

\item{n_boot}{number of bootstrap replicates}

\item{n_boot_try}{max number of attempts for bootstrap resampling}

\item{effect_dir}{direction of beneficial effect, defaults to "positive" for beneficial outcome. Used for one-side tests of bounds.}
}
\value{
list with observed difference between plc and vax, Hudgens-style test statistic, and p-value
}
\description{
Function for bootstrap test of Hudgens-style test statistic
}
