% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_boot_df}
\alias{make_boot_df}
\title{Helper function to make dataframe from bootstrap results}
\usage{
make_boot_df(boot_estimates, estimand = "nat_inf", method = "gcomp")
}
\arguments{
\item{boot_estimates}{list of n_boot bootstrap estimates from one_boot}

\item{estimand}{estimand of interest}

\item{method}{method of interest}
}
\value{
data frame with bootstrap results for point estimates & bounds
}
\description{
Helper function to make dataframe from bootstrap results
}
