% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{one_boot}
\alias{one_boot}
\title{Function for one bootstrap sample}
\usage{
one_boot(
  data,
  G_name = "G",
  V_name = "V",
  X_name = "X",
  Y_name = "Y",
  est = c("gcomp_pop_estimand", "gcomp", "efficient_aipw", "efficient_tmle",
    "hudgens_adj_upper", "hudgens_adj_lower"),
  G_X_model = NULL,
  Y_X_model = NULL,
  family = "gaussian"
)
}
\arguments{
\item{data}{original data to bootstrap}

\item{G_name}{growth outcome variable name}

\item{V_name}{vaccination variable name}

\item{X_name}{covariate name(s)}

\item{Y_name}{infection variable name}

\item{est}{character vector of names of estimators to use for growth effect}

\item{G_X_model}{optional specify model to be used for fitting growth on covariates, otherwise growth on all covariates}

\item{Y_X_model}{optional specify model to be used for fitting infection on covariates, otherwise infection on all covariates}

\item{family}{family for outcome model, defaults to gaussian for growth}
}
\value{
list containing results for specified estimators on single bootstrap sample
}
\description{
Function for one bootstrap sample
}
