% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation_bound_doomed}
\alias{permutation_bound_doomed}
\title{Function for permutation test of bounds in doomed}
\usage{
permutation_bound_doomed(
  data,
  Y_name = "Y",
  Z_name = "Z",
  S_name = "S",
  n_permutations = 1000,
  family = "gaussian",
  effect_dir = "positive"
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{Y_name}{growth outcome variable name}

\item{Z_name}{vaccination variable name}

\item{S_name}{infection variable name}

\item{n_permutations}{number of permutations to run, default 1000}

\item{family}{defaults to gaussian for continuous outcome}

\item{effect_dir}{direction of beneficial effect, defaults to "positive" for beneficial outcome. Used for one-side tests of bounds.}
}
\value{
list with original estimate, null estimates for n_permutations, and p_value for one-sided test based on effect_dir
}
\description{
Function for permutation test of bounds in doomed
}
