% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_tests.R
\name{plot.chop_lump_res}
\alias{plot.chop_lump_res}
\title{Helper function to plot chop-lump test results}
\usage{
\method{plot}{chop_lump_res}(out)
}
\arguments{
\item{out}{chop_lump_res object to plot}
}
\value{
A histogram plot with the distribution of test statistics under the null hypothesis,
and a vertical line showing the observed test statistic.
}
\description{
Helper function to plot chop-lump test results
}
