% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.sens}
\alias{plot.sens}
\title{Plot method for sens objects}
\usage{
\method{plot}{sens}(object, se = TRUE, effect_type = c("additive", "multiplicative"), ...)
}
\arguments{
\item{object}{An object of class "sens"}

\item{se}{Logical; whether to include error bars}

\item{effect_type}{Character; either "additive" or "multiplicative"}

\item{...}{Additional arguments passed to plot (not used here)}
}
\description{
Plot method for sens objects
}
