% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{provide}
\alias{provide}
\title{Example simulated dataset based on the PROVIDE study}
\format{
A data frame with X rows and Y variables:
\describe{
\item{wk10_haz}{Height-for-age Z-score at 10 weeks of age (numeric)}
\item{gender}{Infant gender, either \code{"Male"} or \code{"Female"} (character)}
\item{num_hh_sleep}{Number of people sleeping in the household (integer)}
\item{rotaarm}{Indicator for assignment to rotavirus vaccination arm (0 = control, 1 = vaccine) (integer)}
\item{rotaepi}{Indicator for rotavirus episode during follow-up (0 = no, 1 = yes) (integer)}
\item{any_abx_wk52}{Any antibiotic usage reported through 52 weeks of age (0 = no, 1 = yes) (integer)}
}
}
\usage{
provide
}
\description{
Example simulated dataset based on the PROVIDE study
}
\keyword{datasets}
