% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_predict}
\alias{simple_predict}
\title{Helper function to predict from a model appropriately irrespective of class}
\usage{
simple_predict(model, newdata)
}
\arguments{
\item{model}{A fitted model of class \code{'glm'} or \code{'SuperLearner'}}

\item{newdata}{A \code{data.frame} to obtain predictions on}
}
\value{
Predictions from \code{model} on \code{newdata}
}
\description{
Helper function to predict from a model appropriately irrespective of class
}
