% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegrowth.R
\name{vegrowth}
\alias{vegrowth}
\title{Main function to get growth effect point estimate and standard error}
\usage{
vegrowth(
  data,
  G_name = "G",
  V_name = "V",
  X_name = "X",
  Y_name = "Y",
  est = c("gcomp_pop_estimand", "gcomp", "efficient_aipw", "efficient_tmle", "choplump",
    "hudgens_lower", "hudgens_upper"),
  n_boot = 1000,
  seed = 12345,
  return_se = FALSE,
  G_X_model = NULL,
  Y_X_model = NULL,
  null_hypothesis_value = 0,
  alpha_level = 0.025
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{G_name}{growth outcome variable name}

\item{V_name}{vaccination variable name}

\item{X_name}{covariate name(s)}

\item{Y_name}{infection variable name}

\item{est}{character vector of names of estimators to use for growth effect}

\item{n_boot}{number of bootstrap replicates}

\item{seed}{seet to set for replicability of bootstrap}

\item{return_se}{indicator to return closed form standard error for efficient_aipw or efficient_tmle, default FALSE}

\item{G_X_model}{optional specify model to be used for fitting growth on covariates, otherwise growth on all covariates}

\item{Y_X_model}{optional specify model to be used for fitting infection on covariates, otherwise infection on all covariates}

\item{null_hypothesis_value}{null value for hypothesis test effect for VE and population estimand g-comp, default 0}

\item{alpha_level}{alpha level for hypothesis testing, default 0.025}
}
\value{
List of class \code{vegrowth}
}
\description{
Main function to get growth effect point estimate and standard error
}
