% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nat_inf.R
\name{do_gcomp_nat_inf}
\alias{do_gcomp_nat_inf}
\title{Function for g-computation of counterfactual post-infection outcomes in the
naturally infected principal strata}
\usage{
do_gcomp_nat_inf(
  data,
  models,
  Z_name = NULL,
  X_name = NULL,
  exclusion_restriction = FALSE,
  two_part_model = FALSE
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{models}{A list of fitted models returned from \code{fit_models}}

\item{Z_name}{vaccination variable name}

\item{X_name}{covariate name(s)}

\item{exclusion_restriction}{boolean or vector of boolean (TRUE,FALSE) indicating version of naturally infected estimators with and/or without exclusion restriction assumptions, default FALSE}

\item{two_part_model}{If \code{exclusion_restriction} is \code{TRUE} or \code{"pop"} is included in \code{estimand}, should E(Y | Z, X) be estimated using separate models for E(Y | Z, X, S) and P(S | Z, X) (if \code{TRUE}) or using a single model for E(Y | Z, X) (if \code{FALSE}). Currently, this is only implemented for \code{'gcomp'} estimators}
}
\value{
g-comp estimate of growth effect in the naturally infected strata
}
\description{
Function for g-computation of counterfactual post-infection outcomes in the
naturally infected principal strata
}
