% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_ml_models}
\alias{fit_ml_models}
\title{Function to fit all SuperLearner models needed by various estimators}
\usage{
fit_ml_models(
  data,
  Y_name = "Y",
  Z_name = "Z",
  X_name = c("X"),
  S_name = "S",
  estimand = c("nat_inf", "doomed", "pop"),
  method = c("gcomp", "ipw", "aipw", "tmle", "bound", "sens"),
  exclusion_restriction = c(TRUE, FALSE),
  Y_Z_X_library = c("SL.glm"),
  Y_X_library = c("SL.glm"),
  S_X_library = c("SL.glm"),
  S_Z_X_library = c("SL.glm"),
  Z_X_library = c("SL.mean"),
  family = "gaussian",
  v_folds = 3
)
}
\arguments{
\item{data}{dataset used to fit models}

\item{Y_name}{name of growth outcome variable, default G}

\item{Z_name}{name of vaccine treatment variable, default Z}

\item{X_name}{character vector containing name(s) of covariates, default X}

\item{S_name}{name of infection variable, default Y}

\item{Y_X_library}{specify SuperLearner libraries to be used for fitting growth on covariates. Default GLM}

\item{S_X_library}{specify SuperLearner libraries to be used for fitting infection on covariates. Default GLM}

\item{family}{family for outcome model, defaults to gaussian for growth}

\item{v_folds}{number of cross-validation folds to use in SuperLearner, default 3}

\item{est}{character vector of names of estimators to use for growth effect}
}
\value{
list with models needed for specified estimators
}
\description{
Function to fit all SuperLearner models needed by various estimators
}
