% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Function to fit all models needed by various estimators}
\usage{
fit_models(
  data,
  Y_name = "Y",
  Z_name = "Z",
  X_name = c("X"),
  S_name = "S",
  estimand = c("nat_inf", "doomed", "pop"),
  method = c("gcomp", "ipw", "aipw", "tmle", "bound", "sens"),
  Y_Z_X_model = NULL,
  Y_X_S1_model = NULL,
  Y_X_S0_model = NULL,
  S_X_model = NULL,
  S_Z_X_model = NULL,
  Z_X_model = paste0(Z_name, " ~ 1"),
  family = "gaussian"
)
}
\arguments{
\item{data}{dataset used to fit models}

\item{Y_name}{name of growth outcome variable, default G}

\item{Z_name}{name of vaccine treatment variable, default Z}

\item{X_name}{character vector containing name(s) of covariates, default X}

\item{S_name}{name of infection variable, default Y}

\item{estimand}{character vector with name(s) of estimands of interest; "nat_inf" = naturally infected, "doomed" = doomed, "pop" = marginal/population-level}

\item{S_X_model}{optional specify model to be used for fitting infection on covariates, otherwise infection on all covariates}

\item{family}{family for outcome model, defaults to gaussian for growth}

\item{Y_X_model}{optional specify model to be used for fitting growth on covariates, otherwise growth on all covariates}
}
\value{
list with models needed for specified estimators
}
\description{
Function to fit all models needed by various estimators
}
