% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_tests.R
\name{hudgens_test_doomed}
\alias{hudgens_test_doomed}
\title{Function for bootstrap test of Hudgens-style test statistic - doomed}
\usage{
hudgens_test_doomed(
  data,
  G_name = "G",
  V_name = "V",
  Y_name = "Y",
  lower_bound = TRUE,
  n_boot = 1000,
  n_boot_try = n_boot * 10
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{G_name}{growth outcome variable name}

\item{V_name}{vaccination variable name}

\item{Y_name}{infection variable name}

\item{lower_bound}{A boolean. If TRUE, then adds the smallest growth measures
to the infected vaccinees thereby yielding a lower
bound on the effect of interest. If FALSE, then adds the largest
growth measures to the infected vacccinees thereby yielding an upper
bound on the effect of interest.
@param n_boot number of bootstrap replicates
@param n_boot_try max number of attempts for bootstrap resampling

@returns list with observed difference between plc and vax, Hudgens-style test statistic, and p-value}
}
\description{
Function for bootstrap test of Hudgens-style test statistic - doomed
}
