% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{one_boot}
\alias{one_boot}
\title{Function for one bootstrap sample}
\usage{
one_boot(
  data,
  Y_name = "Y",
  Z_name = "Z",
  X_name = "X",
  S_name = "S",
  estimand = c("nat_inf", "doomed", "pop"),
  method = c("gcomp", "ipw", "aipw", "tmle", "bound", "sens"),
  ml = FALSE,
  Y_Z_X_model = NULL,
  Y_X_S1_model = NULL,
  Y_X_S0_model = NULL,
  S_X_model = NULL,
  S_Z_X_model = NULL,
  Z_X_model = NULL,
  Y_Z_X_library = c("SL.glm"),
  Y_X_library = c("SL.glm"),
  S_X_library = c("SL.glm"),
  S_Z_X_library = c("SL.glm"),
  Z_X_library = c("SL.mean"),
  family = "gaussian",
  v_folds = 3,
  effect_dir = "positive",
  epsilon = exp(seq(log(0.5), log(2), length = 50)),
  max_resample = 10,
  return_se = TRUE
)
}
\arguments{
\item{data}{original data to bootstrap}

\item{Y_name}{growth outcome variable name}

\item{Z_name}{vaccination variable name}

\item{X_name}{covariate name(s)}

\item{S_name}{infection variable name}

\item{ml}{boolean to use SuperLearner models, default FALSE}

\item{Y_X_S1_model}{optional specify model to be used for fitting growth on covariates in the infected, otherwise growth on all covariates}

\item{Y_X_S0_model}{optional specify model to be used for fitting growth on covariates in the uninfected, otherwise growth on all covariates}

\item{S_X_model}{optional specify model to be used for fitting infection on covariates, otherwise infection on all covariates}

\item{Y_X_library}{optional specify SuperLearner libraries for model fitting growth on covariates, default glm}

\item{S_X_library}{optional specify SuperLearner libraries for model fitting infection on covariates, default glm}

\item{family}{family for outcome model, defaults to gaussian for growth}

\item{v_folds}{number of cross validation folds for SuperLearner, default 3}

\item{effect_dir}{direction of beneficial effect, defaults to "positive" for beneficial outcome. Used for one-side tests of bounds.}

\item{epsilon}{a vector of values for the sensitivity parameter}

\item{est}{character vector of names of estimators to use for growth effect}
}
\value{
list containing results for specified estimators on single bootstrap sample
}
\description{
Function for one bootstrap sample
}
