% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.vegrowth}
\alias{print.vegrowth}
\title{Print the output of a \code{"vegrowth"} object}
\usage{
\method{print}{vegrowth}(x, ...)
}
\arguments{
\item{x}{An \code{"vegrowth"} object.}

\item{...}{other arguments (not used)}
}
\description{
Print the output of a \code{"vegrowth"} object
}
