% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegrowth.R
\name{vegrowth}
\alias{vegrowth}
\title{Main function to get growth effect point estimate and standard error}
\usage{
vegrowth(
  data,
  G_name = "G",
  V_name = "V",
  X_name = "X",
  Y_name = "Y",
  est = c("gcomp_pop_estimand", "gcomp", "efficient_aipw", "efficient_tmle", "choplump",
    "hudgens_lower", "hudgens_upper", "hudgens_adj_lower", "hudgens_adj_upper"),
  n_boot = 1000,
  seed = 12345,
  return_se = FALSE,
  ml = FALSE,
  G_V_X_model = NULL,
  G_X_Y1_model = NULL,
  G_X_Y0_model = NULL,
  Y_X_model = NULL,
  G_V_X_library = c("SL.glm"),
  G_X_library = c("SL.glm"),
  Y_X_library = c("SL.glm"),
  null_hypothesis_value = 0,
  alpha_level = 0.025,
  return_models = TRUE,
  family = "gaussian",
  v_folds = 3,
  effect_dir = "positive"
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{G_name}{growth outcome variable name}

\item{V_name}{vaccination variable name}

\item{X_name}{covariate name(s)}

\item{Y_name}{infection variable name}

\item{est}{character vector of names of estimators to use for growth effect}

\item{n_boot}{number of bootstrap replicates}

\item{seed}{seet to set for replicability of bootstrap}

\item{return_se}{indicator to return closed form standard error for efficient_aipw or efficient_tmle, default FALSE}

\item{ml}{boolean to use SuperLearner models, default FALSE}

\item{G_V_X_model}{optional specify model to be used for fitting growth on vaccine + covariates, otherwise growth on all covariates}

\item{G_X_Y1_model}{optional specify model to be used for fitting growth on covariates in the infected, otherwise growth on all covariates}

\item{G_X_Y0_model}{optional specify model to be used for fitting growth on covariates in the uninfected, otherwise growth on all covariates}

\item{Y_X_model}{optional specify model to be used for fitting infection on covariates, otherwise infection on all covariates}

\item{G_V_X_library}{optional specify SuperLearner libraries for model fitting growth on covariates + vaccine, default glm}

\item{G_X_library}{optional specify SuperLearner libraries for model fitting growth on covariates, default glm}

\item{Y_X_library}{optional specify SuperLearner libraries for model fitting infection on covariates, default glm}

\item{null_hypothesis_value}{null value for hypothesis test effect for VE and population estimand g-comp, default 0}

\item{alpha_level}{alpha level for hypothesis testing, default 0.025}

\item{return_models}{boolean return models, default TRUE}

\item{family}{family for outcome variable 'G', defaults to gaussian for growth}

\item{v_folds}{number of cross validation folds for SuperLearner, default 3}

\item{effect_dir}{direction of beneficial effect, defaults to "positive" for beneficial outcome. Used for one-side tests of bounds.}
}
\value{
List of class \code{vegrowth}
}
\description{
Main function to get growth effect point estimate and standard error
}
