% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegrowth.R
\name{vegrowth}
\alias{vegrowth}
\title{Main function to get growth effect point estimate and standard error}
\usage{
vegrowth(
  data,
  Y_name = "G",
  Z_name = "Z",
  X_name = "X",
  S_name = "S",
  estimand = c("nat_inf", "doomed", "pop"),
  method = c("gcomp", "ipw", "aipw", "tmle", "bound", "sens"),
  n_boot = 1000,
  permutation = FALSE,
  n_perm = 1000,
  seed = 12345,
  return_se = TRUE,
  ml = TRUE,
  Y_Z_X_model = NULL,
  Y_X_S1_model = NULL,
  Y_X_S0_model = NULL,
  S_X_model = NULL,
  S_Z_X_model = NULL,
  Z_X_model = paste0(Z_name, " ~ 1"),
  Y_Z_X_library = c("SL.glm"),
  Y_X_library = c("SL.glm"),
  S_X_library = c("SL.glm"),
  S_Z_X_library = c("SL.glm"),
  Z_X_library = c("SL.mean"),
  null_hypothesis_value = 0,
  alpha_level = 0.05,
  return_models = TRUE,
  family = "gaussian",
  v_folds = 3,
  effect_dir = "positive",
  epsilon = exp(seq(log(0.5), log(2), length = 49))
)
}
\arguments{
\item{data}{dataframe containing dataset to use for analysis}

\item{Y_name}{growth outcome variable name}

\item{Z_name}{vaccination variable name}

\item{X_name}{covariate name(s)}

\item{S_name}{infection variable name}

\item{estimand}{character vector with name(s) of estimands of interest; "nat_inf" = naturally infected, "doomed" = doomed, "pop" = marginal/population-level}

\item{method}{character vector with name(s) of methods to use for estimation; "gcomp" = g-computation, "ipw" = inverese probability weighting, "aipw" = augmented inverse probability weighting, "tmle" = targeted maximum likelihood estimation (nat_inf only), "bound" = bounds without cross-world assumptions (nat_inf and doomed only), "sens" = sensitivity analysis (nat_inf only)}

\item{n_boot}{number of bootstrap replicates}

\item{seed}{seet to set for replicability of bootstrap}

\item{return_se}{indicator to return closed form standard error for efficient_aipw or efficient_tmle, default TRUE}

\item{ml}{boolean to use SuperLearner models for AIPW & TMLE, default TRUE}

\item{Y_Z_X_model}{optional specify model to be used for fitting growth on vaccine + covariates, otherwise growth on all covariates}

\item{Y_X_S1_model}{optional specify model to be used for fitting growth on covariates in the infected, otherwise growth on all covariates}

\item{Y_X_S0_model}{optional specify model to be used for fitting growth on covariates in the uninfected, otherwise growth on all covariates}

\item{S_X_model}{optional specify model to be used for fitting infection on covariates, otherwise infection on all covariates}

\item{Y_Z_X_library}{optional specify SuperLearner libraries for model fitting growth on covariates + vaccine, default glm}

\item{Y_X_library}{optional specify SuperLearner libraries for model fitting growth on covariates, default glm}

\item{S_X_library}{optional specify SuperLearner libraries for model fitting infection on covariates, default glm}

\item{null_hypothesis_value}{null value for hypothesis test effect for ZE and population estimand g-comp, default 0}

\item{alpha_level}{alpha level for hypothesis testing, default 0.025}

\item{return_models}{boolean return models, default TRUE}

\item{family}{family for outcome variable 'G', defaults to gaussian for growth}

\item{v_folds}{number of cross validation folds for SuperLearner, default 3}

\item{effect_dir}{direction of beneficial effect, defaults to "positive" for beneficial outcome. Used for one-side tests of bounds.}

\item{epsilon}{a vector of values for the sensitivity parameter (only applicable for method = "sens")}
}
\value{
List of class \code{vegrowth}
}
\description{
Main function to get growth effect point estimate and standard error
}
